% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedTools.R
\name{intersectAndExtendBed}
\alias{intersectAndExtendBed}
\title{Intersect and Extend regions from multiple bed files}
\usage{
intersectAndExtendBed(bedFiles, minOverlap)
}
\arguments{
\item{bedFiles}{A character vector of bed files}

\item{minOverlap}{Min number of files or fraction of file to have a region}
}
\value{
A granges oject of the merged regions
}
\description{
intersectAndExtendBed is a function which will intersect a given vector
of bed files and return a single Granges object of the merged bed file
}
\examples{
\dontrun{
intersectAndExtendBed(bedFiles = c("/path/to/A.bed", "/path/to/B.bed", "/path/to/C.bed"),
minOverlap = 2)
}

}
